Explicit casts to (off_t), (size_t) in multitape_read.c

+++ tar/multitape/multitape_read.c
@@ -124,12 +124,12 @@ stream_get_chunk(struct stream * S, const uint8_t ** buf, size_t * clen,
 
 	/* Skip part of the current chunk if appropriate. */
 	if (S->skiplen) {
-		skip = S->chunklen - S->chunkpos;
+		skip = (off_t)(S->chunklen - S->chunkpos);
 		if (skip > S->skiplen)
 			skip = S->skiplen;
 
 		S->skiplen -= skip;
-		S->chunkpos += skip;
+		S->chunkpos += (size_t)skip;
 	}
 
 	while ((S->chunklen == S->chunkpos) && (S->istr != NULL)) {
@@ -151,7 +151,7 @@ stream_get_chunk(struct stream * S, const uint8_t ** buf, size_t * clen,
 		S->chunklen = len;
 
 		/* Set current position within buffer. */
-		S->chunkpos = S->skiplen;
+		S->chunkpos = (size_t)S->skiplen;
 		S->skiplen = 0;
 
 		/* The chunk is no longer pending. */
@@ -374,7 +374,7 @@ readtape_read(TAPE_R * d, const void ** buffer)
 		if (stream_get_chunk(readstream, buf, &clen, d->C))
 			goto err0;
 		if ((off_t)clen > *readmaxlen)
-			clen = *readmaxlen;
+			clen = (size_t)(*readmaxlen);
 
 		readstream->chunkpos += clen;
 		*readmaxlen -= clen;
